requirements.txt
# Core Dependencies for THE_DATAZENtr

# LangChain & LangSmith
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langsmith>=0.0.70
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5

# Memory & Database
asyncpg>=0.29.0
pgvector>=0.2.4
redis>=5.0.1
sqlite3

# AI/ML
anthropic>=0.18.1
openai>=1.12.0
numpy>=1.24.3
tiktoken>=0.5.2

# Web Framework (for webapp)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.26.0
rich>=13.7.0

# Development & Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=23.12.1
ruff>=0.1.11

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.3